Add wallet test added
