+++ test/e2e/tests/multichain-accounts/account-details.spec.ts
@@ -1,9 +1,15 @@
 import { Suite } from 'mocha';
+import { withFixtures, WALLET_PASSWORD } from '../../helpers';
+import FixtureBuilder from '../../fixture-builder';
 import AccountListPage from '../../page-objects/pages/account-list-page';
+import HeaderNavbar from '../../page-objects/pages/header-navbar';
+import AccountDetailsModal from '../../page-objects/pages/dialog/account-details-modal';
 import AccountAddressModal from '../../page-objects/pages/multichain/account-address-modal';
 import AddressListModal from '../../page-objects/pages/multichain/address-list-modal';
 import MultichainAccountDetailsPage from '../../page-objects/pages/multichain/multichain-account-details-page';
 import MultichainWalletDetailsPage from '../../page-objects/pages/multichain/multichain-wallet-details-page';
+import PrivateKeyModal from '../../page-objects/pages/multichain/private-key-modal';
+import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
 import { Driver } from '../../webdriver/driver';
 import { withMultichainAccountsDesignEnabled } from './common';
 
@@ -80,6 +86,63 @@ describe('Multichain Accounts - Account Details', function (this: Suite) {
               'Wallet row should be present and contain wallet name',
             );
           }
+        },
+      );
+    });
+  });
+
+  describe('Show account details', function () {
+    it('should show the correct private key from account menu', async function () {
+      await withMultichainAccountsDesignEnabled(
+        {
+          title: this.test?.fullTitle(),
+        },
+        async (driver: Driver) => {
+          const accountListPage = new AccountListPage(driver);
+          await accountListPage.checkPageIsLoaded({
+            isMultichainAccountsState2Enabled: true,
+          });
+          await accountListPage.openMultichainAccountMenu({
+            accountLabel: 'Account 1',
+          });
+          await accountListPage.clickMultichainAccountMenuItem(
+            'Account details',
+          );
+
+          const accountDetailsPage = new MultichainAccountDetailsPage(driver);
+          await accountDetailsPage.clickPrivateKeyRow();
+
+          const privateKeyModal = new PrivateKeyModal(driver);
+          await privateKeyModal.checkPageIsLoaded();
+          await privateKeyModal.typePassword(WALLET_PASSWORD);
+          await privateKeyModal.clickConfirm();
+          const accountDetailsModal = new AccountDetailsModal(driver);
+          await accountDetailsModal.clickCopyPrivateKeyButton();
+          await accountDetailsModal.checkAddressIsCopied();
+        },
+      );
+    });
+
+    it('should show the correct private key from global menu', async function () {
+      await withFixtures(
+        {
+          fixtures: new FixtureBuilder().build(),
+          title: this.test?.fullTitle(),
+          forceBip44Version: 2,
+        },
+        async ({ driver }) => {
+          await loginWithoutBalanceValidation(driver);
+          const headerNavbar = new HeaderNavbar(driver);
+          await headerNavbar.openAccountDetailsModal();
+          const accountDetailsPage = new MultichainAccountDetailsPage(driver);
+          await accountDetailsPage.clickPrivateKeyRow();
+          const privateKeyModal = new PrivateKeyModal(driver);
+          await privateKeyModal.checkPageIsLoaded();
+          await privateKeyModal.typePassword(WALLET_PASSWORD);
+          await privateKeyModal.clickConfirm();
+          const accountDetailsModal = new AccountDetailsModal(driver);
+          await accountDetailsModal.clickCopyPrivateKeyButton();
+          await accountDetailsModal.checkAddressIsCopied();
         },
       );
     });
