+++ test/e2e/tests/multichain-accounts/add-account.spec.ts
@@ -10,13 +10,23 @@ import ResetPasswordPage from '../../page-objects/pages/reset-password-page';
 import MultichainAccountDetailsPage from '../../page-objects/pages/multichain/multichain-account-details-page';
 import { Driver } from '../../webdriver/driver';
 import { loginWithoutBalanceValidation } from '../../page-objects/flows/login.flow';
-import { withMultichainAccountsDesignEnabled } from './common';
+import {
+  withImportedAccount,
+  withMultichainAccountsDesignEnabled,
+} from './common';
+
+const SECOND_ACCOUNT_NAME = 'Account 2';
+const IMPORTED_ACCOUNT_NAME = 'Imported Account 1';
+const CUSTOM_ACCOUNT_NAME = 'Custom 1';
+const TEST_PRIVATE_KEY =
+  '14abe6f4aab7f9f626fe981c864d0adeb5685f289ac9270c27b8fd790b4235d6';
+
+const importedAccount = {
+  name: 'Imported Account 1',
+  address: '0x7A46ce51fbBB29C34aea1fE9833c27b5D2781925',
+};
 
 describe('Add account', function () {
-  const SECOND_ACCOUNT_NAME = 'Account 2';
-  const IMPORTED_ACCOUNT_NAME = 'Imported Account 1';
-  const CUSTOM_ACCOUNT_NAME = 'Custom 1';
-
   it('should not affect public address when using secret recovery phrase to recover account with non-zero balance', async function () {
     await withMultichainAccountsDesignEnabled(
       {
@@ -71,6 +81,36 @@ describe('Add account', function () {
         await headerNavbar.checkAccountLabel(SECOND_ACCOUNT_NAME);
         // BUG 37030 With BIP44 enabled wallet is not showing balance
         // await homePage.checkExpectedBalanceIsDisplayed('2.8');
+      },
+    );
+  });
+
+  it('should remove imported private key account successfully', async function () {
+    await withImportedAccount(
+      {
+        title: this.test?.fullTitle(),
+        privateKey: TEST_PRIVATE_KEY,
+      },
+      async (driver: Driver) => {
+        const accountListPage = new AccountListPage(driver);
+        await accountListPage.checkPageIsLoaded({
+          isMultichainAccountsState2Enabled: true,
+        });
+        await accountListPage.openMultichainAccountMenu({
+          accountLabel: importedAccount.name,
+        });
+        await accountListPage.clickMultichainAccountMenuItem('Account details');
+
+        const accountDetailsPage = new MultichainAccountDetailsPage(driver);
+        await accountDetailsPage.checkPageIsLoaded();
+
+        await accountDetailsPage.clickRemoveAccountButton();
+
+        await accountDetailsPage.clickRemoveAccountConfirmButton();
+
+        await accountListPage.checkAccountIsNotDisplayedInAccountList(
+          importedAccount.name,
+        );
       },
     );
   });
@@ -138,7 +178,7 @@ describe('Add account', function () {
     );
   });
 
-  it('persists custom account label through account change and wallet lock', async function () {
+  it('added account should persiste after wallet lock', async function () {
     await withMultichainAccountsDesignEnabled(
       {
         title: this.test?.fullTitle(),
